Added support for parsing JSON 1.0 format.
According to
https://nvd.nist.gov/General/News/XML-Vulnerability-Feed-Retirement,
NIST has retired the XML version of the data feeds.  Upstream does not
have a published fix to move over to JSON, but a user did supply a pull
request that tried to add JSON 1.0 parsing support.  I applied the
content of the patch, but also kept the older XML parsing under the USE_XML
CFLAG.  Since it is not defined, the code will build the JSON parsing
instead.  This code required the jansson library.  I had to build it
from source, since it was not avaiable a a package for gentoo/sabayon.
It had to be installed with PREFIX=/usr to make building cve-check-tool
happy (libtool).  I tried to pull and parse JSON 1.1 format feeds, but
the JSON format is different and will need updated parsers.  The URLs
are still in the code, but I did not start trying to update the parsers.

Upstream JSON Patch:
https://patch-diff.githubusercontent.com/raw/clearlinux/cve-check-tool/pull/74.diff

jansson source:
https://github.com/akheron/jansson